move main functions into pre_execution
