+++ lib/internal/modules/cjs/loader.js
@@ -73,7 +73,7 @@ const {
   resolveMainPath,
   shouldUseESMLoader,
   runMainESM
-} = require('internal/process/main');
+} = require('internal/bootstrap/pre_execution');
 const pendingDeprecation = getOptionValue('--pending-deprecation');
 
 module.exports = { wrapSafe, Module, toRealPath, readPackageScope };
