+++ test/parallel/test-bootstrap-modules.js
@@ -51,7 +51,6 @@ const expectedModules = new Set([
   'NativeModule internal/options',
   'NativeModule internal/priority_queue',
   'NativeModule internal/process/execution',
-  'NativeModule internal/process/main',
   'NativeModule internal/process/per_thread',
   'NativeModule internal/process/promises',
   'NativeModule internal/process/task_queues',
